--- libs/common/src/billing/abstractions/provider-billing.service.abstraction.ts
@@ -1,11 +1,0 @@
-import { PlanType } from "../enums";
-
-export abstract class ProviderBillingServiceAbstraction {
-  createClientOrganization: (
-    providerId: string,
-    name: string,
-    ownerEmail: string,
-    planType: PlanType,
-    seats: number,
-  ) => Promise<void>;
-}
