+++ libs/common/src/billing/services/organization-billing.service.ts
@@ -3,11 +3,12 @@ import { OrganizationApiServiceAbstraction as OrganizationApiService } from "../
 import { OrganizationCreateRequest } from "../../admin-console/models/request/organization-create.request";
 import { OrganizationKeysRequest } from "../../admin-console/models/request/organization-keys.request";
 import { OrganizationResponse } from "../../admin-console/models/response/organization.response";
+import { CryptoService } from "../../platform/abstractions/crypto.service";
+import { EncryptService } from "../../platform/abstractions/encrypt.service";
+import { I18nService } from "../../platform/abstractions/i18n.service";
+import { EncString } from "../../platform/models/domain/enc-string";
+import { OrgKey } from "../../types/key";
 import { SyncService } from "../../vault/abstractions/sync/sync.service.abstraction";
-import {
-  BillingKeyGenerationServiceAbstraction,
-  GeneratedKeys,
-} from "../abstractions/billing-key-generation-service.abstraction";
 import {
   OrganizationBillingServiceAbstraction,
   OrganizationInformation,
@@ -17,10 +18,19 @@ import {
 } from "../abstractions/organization-billing.service";
 import { PlanType } from "../enums";
 
+interface OrganizationKeys {
+  encryptedKey: EncString;
+  publicKey: string;
+  encryptedPrivateKey: EncString;
+  encryptedCollectionName: EncString;
+}
+
 export class OrganizationBillingService implements OrganizationBillingServiceAbstraction {
   constructor(
-    private billingKeyGenerationService: BillingKeyGenerationServiceAbstraction,
     private apiService: ApiService,
+    private cryptoService: CryptoService,
+    private encryptService: EncryptService,
+    private i18nService: I18nService,
     private organizationApiService: OrganizationApiService,
     private syncService: SyncService,
   ) {}
@@ -28,7 +38,7 @@ export class OrganizationBillingService implements OrganizationBillingServiceAbs
   async purchaseSubscription(subscription: SubscriptionInformation): Promise<OrganizationResponse> {
     const request = new OrganizationCreateRequest();
 
-    const organizationKeys = await this.billingKeyGenerationService.generateOrganizationKeys();
+    const organizationKeys = await this.makeOrganizationKeys();
 
     this.setOrganizationKeys(request, organizationKeys);
 
@@ -38,24 +48,6 @@ export class OrganizationBillingService implements OrganizationBillingServiceAbs
 
     this.setPaymentInformation(request, subscription.payment);
 
-    return await this.createOrganization(request);
-  }
-
-  async startFree(subscription: SubscriptionInformation): Promise<OrganizationResponse> {
-    const request = new OrganizationCreateRequest();
-
-    const organizationKeys = await this.billingKeyGenerationService.generateOrganizationKeys();
-
-    this.setOrganizationKeys(request, organizationKeys);
-
-    this.setOrganizationInformation(request, subscription.organization);
-
-    this.setPlanInformation(request, subscription.plan);
-
-    return await this.createOrganization(request);
-  }
-
-  private async createOrganization(request: OrganizationCreateRequest) {
     const response = await this.organizationApiService.create(request);
 
     await this.apiService.refreshIdentityToken();
@@ -63,6 +55,41 @@ export class OrganizationBillingService implements OrganizationBillingServiceAbs
     await this.syncService.fullSync(true);
 
     return response;
+  }
+
+  async startFree(subscription: SubscriptionInformation): Promise<OrganizationResponse> {
+    const request = new OrganizationCreateRequest();
+
+    const organizationKeys = await this.makeOrganizationKeys();
+
+    this.setOrganizationKeys(request, organizationKeys);
+
+    this.setOrganizationInformation(request, subscription.organization);
+
+    this.setPlanInformation(request, subscription.plan);
+
+    const response = await this.organizationApiService.create(request);
+
+    await this.apiService.refreshIdentityToken();
+
+    await this.syncService.fullSync(true);
+
+    return response;
+  }
+
+  private async makeOrganizationKeys(): Promise<OrganizationKeys> {
+    const [encryptedKey, key] = await this.cryptoService.makeOrgKey<OrgKey>();
+    const [publicKey, encryptedPrivateKey] = await this.cryptoService.makeKeyPair(key);
+    const encryptedCollectionName = await this.encryptService.encrypt(
+      this.i18nService.t("defaultCollection"),
+      key,
+    );
+    return {
+      encryptedKey,
+      publicKey,
+      encryptedPrivateKey,
+      encryptedCollectionName,
+    };
   }
 
   private prohibitsAdditionalSeats(planType: PlanType) {
@@ -88,7 +115,7 @@ export class OrganizationBillingService implements OrganizationBillingServiceAbs
     request.initiationPath = information.initiationPath;
   }
 
-  private setOrganizationKeys(request: OrganizationCreateRequest, keys: GeneratedKeys): void {
+  private setOrganizationKeys(request: OrganizationCreateRequest, keys: OrganizationKeys): void {
     request.key = keys.encryptedKey.encryptedString;
     request.keys = new OrganizationKeysRequest(
       keys.publicKey,
